Address code review feedback: fix date formatting, improve code clarity, add backward compatibility

+++ frontend/src/components/Testimonials/Testimonials.js
@@ -46,7 +46,7 @@ const Testimonials = () => {
 
   const formatDate = (dateString) => {
     const date = new Date(dateString);
-    return date.toLocaleDateString('en-US', { month: 'numeric', day: 'numeric', year: 'numeric' });
+    return date.toLocaleDateString('en-US', { month: 'short', day: 'numeric', year: 'numeric' });
   };
 
   const getUserTypeDisplay = (role) => {
